requirements.txt
fastapi==0.115.0
uvicorn==0.30.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.5.1
rank-bm25==0.2.2
PyMuPDF==1.24.10
python-dotenv==1.0.1
